Added exception classes. Refactored most of decode and execute stages, and started on write stage.

+++ src/Component.cpp
@@ -14,7 +14,7 @@ Component::Component(sc_module_name& name) : sc_module(name), id(-1) {
 
 /* NOTE: if this constructor is used, end_module() must be placed at
  *       the end of the constructor this is called from. */
-Component::Component(sc_module_name& name, int ID)
+Component::Component(sc_module_name& name, uint16_t ID)
     : sc_module(makeName(name, ID)), id(ID) {
 
 }
